+++ client/src/components/layout/TopBar.js
@@ -1,7 +1,7 @@
 import React from "react"
 import { Link } from "react-router-dom"
-import SignOutButton from "../authentication/SignOutButton"
 
+import SignOutButton from "../authentication/SignOutButton"
 import Logo from "../../assets/logo.png"
 
 const TopBar = ({ user }) => {
