+++ server/src/routes/api/v1/statsRouter.js
@@ -14,7 +14,7 @@ statsRouter.get("/playerId=:playerId&season=:season", (req, res) => {
     } else {
       const parsedResponse = JSON.parse(data)
       const stats = StatsSerializer.getSummary(parsedResponse.data[0])
-      res.set({ "Content-Type": "application/json" }).status(200).json(stats)
+      return res.status(200).json(stats)
     }
   })
 })
